# perf(http): avoid clone getting request method and url