Code:

```js
Deno.serve({ port: 8085 }, request => {
  return new Response(request.url);
});
```

Before:

```
$ wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    65.49us   36.05us   5.46ms   96.33%
    Req/Sec    75.71k     1.90k   79.44k    78.87%
  9053462 requests in 1.00m, 1.36GB read
Requests/sec: 150641.83
Transfer/sec:     23.13MB
```

After:

```
$ wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    64.72us   38.14us   3.29ms   95.97%
    Req/Sec    76.66k     1.85k   80.17k    77.45%
  9168694 requests in 1.00m, 1.37GB read
Requests/sec: 152557.21
Transfer/sec:     23.42MB
```